# make prod faster for categorical