@wouterwln 2025-06-24
@bvdmitri can we merge this? The failing test seems unrelated to this PR and it saves me about 50% of allocations and time in multiplying categoricals (in experiments as well)